+++ packages/openapi/package.json
@@ -8,10 +8,16 @@
   "types": "./src/index.ts",
   "exports": {
     ".": {
-      "types": "./src/index.ts"
+      "types": "./src/index.ts",
+      "import": "./src/index.ts"
+    },
+    "./client": {
+      "types": "./src/client.ts",
+      "import": "./src/client.ts"
     },
     "./openapi-paths": {
-      "types": "./src/openapi-paths.ts"
+      "types": "./src/openapi-paths.ts",
+      "import": "./src/openapi-paths.ts"
     },
     "./openapi.json": "./openapi.json"
   },
@@ -19,6 +25,11 @@
     "check": "bun run typecheck",
     "generate": "bun ../../scripts/write-openapi.ts && openapi-typescript openapi.json -o src/openapi-paths.ts",
     "typecheck": "tsc --noEmit -p tsconfig.json"
+  },
+  "dependencies": {
+    "@effect/schema": "^0.75.5",
+    "effect": "^3.21.3",
+    "openapi-fetch": "^0.17.0"
   },
   "devDependencies": {
     "openapi-typescript": "^7.13.0",
