+++ packages/openapi/tsconfig.json
@@ -2,6 +2,7 @@
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "rootDir": ".",
+    "lib": ["ES2023", "DOM"],
     "types": []
   },
   "include": ["src/**/*"],
